+++ src/modules/process-ifc/entities/ref-file.entity.ts
@@ -1,0 +1,7 @@
+export class RefFileEntity {
+  constructor(
+    public readonly id: number,
+    public readonly nomFile: string,
+    public readonly pathStorage: string,
+  ) {}
+}