+++ src/modules/process-ifc/interfaces/processifc.interfaces.ts
@@ -1,0 +1,11 @@
+import { RefFileEntity } from '../entities/ref-file.entity';
+
+export type CreateRefFileData = {
+  rutaCompleta: string;
+  nombreArchivo: string;
+};
+
+export abstract class ProcessIfcRepository {
+  abstract create(data: CreateRefFileData): Promise<RefFileEntity>;
+  abstract findAll(): Promise<RefFileEntity[]>;
+}